add roles and permissions and change finacial reports observer class

+++ resources/views/livewire/table-actions.blade.php
@@ -1,10 +1,12 @@
-@if($table === 'permission-form' || $table === 'role-form' )
+@can($viewPermisson)
+
+@if($modal === 'permission-form' || $modal === 'role-form' || $modal === 'user-form' )
     <flux:button 
         icon="eye"
         variant="primary"
         color="cyan"
         class="cursor-pointer"
-        wire:click="$dispatch('view-table', { table: {{ json_encode($table ?? 'unknown') }} , row: {{ json_encode($row) }} })"
+        wire:click="$dispatch('view-table', { modal: {{ json_encode($modal ?? 'unknown') }} , row: {{ json_encode($row) }} })"
     >
     </flux:button>
 @else
@@ -18,17 +20,24 @@
     </flux:button>
 @endif
 
+@endcan
+
+@can($editPermisson)
+
     <flux:button icon="pencil"
-     wire:click="$dispatch('edit-table', { table: {{ json_encode($table ?? 'unknown') }} , row: {{ json_encode($row) }} })"
+     wire:click="$dispatch('edit-table', { modal: {{ json_encode($modal ?? 'unknown') }} , row: {{ json_encode($row) }} })"
       variant="primary" color="cyan" class="cursor-pointer"></flux:button>
+
+    @endcan
 
   <!-- <flux:button icon="trash"
   wire:click="$dispatch('delete-table', { table: {{ json_encode($table ?? 'unknown') }} , row: {{ json_encode($row) }} })"
   variant="primary" color="rose" class="cursor-pointer">Delete</flux:button>  -->
 
   <!-- Delete Button – NOW PERFECTLY ALIGNED & SAFE -->
+   @can($deletePermisson)
 <flux:button icon="trash"
-             wire:click="$dispatch('delete-table', { table: {{ json_encode($table ?? 'unknown') }} , row: {{ json_encode($row) }} })"
+             wire:click="$dispatch('delete-table', { modal: {{ json_encode($modal ?? 'unknown') }} , row: {{ json_encode($row) }} })"
              wire:loading.remove.delay.short
              wire:loading.attr="disabled"
              variant="primary" 
@@ -37,3 +46,4 @@
              wire:confirm="Delete this row permanently?">
 
 </flux:button>
+@endcan
